requirements.txt
streamlit==1.49.1
python-dotenv==1.0.1
pypdf==6.1.0

langchain==0.3.27
langchain-community==0.3.29
langchain-openai==0.3.33
langchain-text-splitters==0.3.9   # 

faiss-cpu==1.8.0.post1  